requirements.txt
python-telegram-bot==22.6
telegramify-markdown>=0.5.4
python-dotenv>=1.0.0
pydantic>=2.0
